requirements.txt
google-generativeai
edge-tts
openai-whisper
moviepy
python-dotenv
imageio-ffmpeg
requests
retry
google-auth-oauthlib
google-api-python-client
openai